+++ brew_view/static/src/js/run.js
@@ -48,28 +48,6 @@ export function appRun($rootScope, $state, $stateParams, $cookies, $http,
     'slate': false,
   };
 
-  $rootScope.login = function() {
-    UtilityService.login().then(function(response) {
-      let token = response.data.token;
-
-      // Save the token to session storage in case we need it later
-      localStorageService.set('token', token);
-
-      // Use the token for all subsequent requests
-      $http.defaults.headers.common.Authorization = 'Bearer ' + token;
-
-      // Now grab the user id and roles from the token
-      let decoded = jwtDecode(token);
-      let userId = decoded.sub;
-
-      // Finally, grab the user definition from the API so we can read prefs
-      $http.get('api/v1/users/' + userId).then(function(response) {
-        $rootScope.userName = response.data.username;
-        $rootScope.changeTheme(response.data.preferences.theme || 'default');
-      });
-    });
-  };
-
   $rootScope.logout = function() {
     localStorageService.remove('token');
     $http.defaults.headers.common.Authorization = undefined;
