+++ brew_view/static/src/partials/login.html
@@ -1,0 +1,7 @@
+<h1 class='page-header'>Login</h1>
+
+<form>
+  <label>Username: <input type="text" ng-model="model.username" /></label><br>
+  <label>Password: <input type="password" ng-model="model.password" /></label><br>
+  <input type="button" ng-click="doSubmit(model)" value="Login">
+</form>
